// src/common/api/model/Tag.ts
/**
 * Profitelo API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface Tag {
    "categoryId"?: string;
    "name": string;
    "id": string;
    "persisted"?: boolean;
    "status": Tag.StatusEnum;
}

export namespace Tag {
    export enum StatusEnum {
        NEW = <any> 'NEW',
        ACCEPTED = <any> 'ACCEPTED',
        REJECTED = <any> 'REJECTED'
    }
}