// src/common/api/model/GetCallDetails.ts
/**
 * Profitelo API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface GetCallDetails {
    "recommendedTags": Array<models.Tag>;
    "serviceOwnerProfile": models.GetProfile;
    "isRecommendable": boolean;
    "isRecommended": boolean;
    "service": models.GetService;
    "serviceUsageDetails": models.GetServiceUsageDetails;
    "expertProfile": models.GetProfile;
}